Get comparison format

Convert a data.frame of all sister groups (from sis_get_sisters) and a vector of 0 and 1 (with names equal to taxon names) to a data.frame with the sister groups that differ in traits.

sis_format_comparison(sisters, trait, phy)

Arguments

sisters

Data.frame from sis_get_sisters()

trait

vector of 0/1 data

phy

A phylo object

Value

data.frame where each row is a sister group comparison.

Examples

data(geospiza, package="geiger")
cleaned <- sis_clean(geospiza$phy, geospiza$dat)
phy <- cleaned$phy
traits <- cleaned$traits
trait <- sis_discretize(traits[,1])
sisters <- sis_get_sisters(phy, ncores=2)
sisters_comparison <- sis_format_comparison(sisters, trait, phy)
print(sisters_comparison)
#>    node         left       right ntax.left ntax.right ntax.total
#> 9    22            1           2         1          1          2
#> 12   25            8           9         1          1          2
#> 8    21         1, 2           3         2          1          3
#> 11   24         8, 9          10         2          1          3
#> 7    20      1, 2, 3           4         3          1          4
#> 10   23            7    8, 9, 10         1          3          4
#> 6    19   1, 2, 3, 4           5         4          1          5
#> 5    18 1, 2, 3,....           6         5          1          6
#> 4    17 1, 2, 3,.... 7, 8, 9, 10         6          4         10
#> 3    16 1, 2, 3,....          11        10          1         11
#> 2    15 1, 2, 3,....          12        11          1         12
#> 1    14 1, 2, 3,....          13        12          1         13
#>                            left.trait right.trait left.unique right.unique
#> 9                                   0           1           0            1
#> 12                                  0           0           0            0
#> 8                                0, 1           1          NA            1
#> 11                               0, 0           0           0            0
#> 7                             0, 1, 1           1          NA            1
#> 10                                  1     0, 0, 0           1            0
#> 6                          0, 1, 1, 1           1          NA            1
#> 5                       0, 1, 1, 1, 1           0          NA            0
#> 4                    0, 1, 1, 1, 1, 0  1, 0, 0, 0          NA           NA
#> 3        0, 1, 1, 1, 1, 0, 1, 0, 0, 0           1          NA            1
#> 2     0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1           0          NA            0
#> 1  0, 1, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0           0          NA            0
#>    ntax.trait0 ntax.trait1
#> 9            1           1
#> 12          NA          NA
#> 8           NA          NA
#> 11          NA          NA
#> 7           NA          NA
#> 10           3           1
#> 6           NA          NA
#> 5           NA          NA
#> 4           NA          NA
#> 3           NA          NA
#> 2           NA          NA
#> 1           NA          NA