Compute multiple tests based on sister group comparisons

Compute multiple tests based on sister group comparisons

sis_test(pairs, drop_matches = TRUE, warn = TRUE)

Arguments

pairs

Data.frame with one row per sister group comparison, with one column for number of taxa in state 0, and one column for the number of taxa in state 1.

drop_matches

Drop sister group comparisons with equal numbers of taxa

warn

Some tests will fail with warnings (too few sister groups or other reasons). Setting this to FALSE will suppress those

Value

A vector with the results of many tests, as well as summary data for the comparisons

Examples

data(geospiza, package="geiger")
cleaned <- sis_clean(geospiza$phy, geospiza$dat)
phy <- cleaned$phy
traits <- cleaned$traits
trait <- sis_discretize(traits[,1])
sisters <- sis_get_sisters(phy)
sisters_comparison <- sis_format_comparison(sisters, trait, phy)
pairs <- sis_format_simpified(sisters_comparison)
sis_test(pairs)
#>       number.comparisons.trait0.bigger       number.comparisons.trait1.bigger 
#>                              1.0000000                              0.0000000 
#> number.comparisons.trait0.equal.trait1        median.proportion.in.state.zero 
#>                              1.0000000                              0.7500000 
#>        median.ntax.diff.zero.minus.one                       pvalue.sign.test 
#>                              2.0000000                              1.0000000 
#>     pvalue.diversity.contrast.ratiolog   pvalue.diversity.contrast.proportion 
#>                              1.0000000                              1.0000000 
#>   pvalue.diversity.contrast.difference     pvalue.diversity.contrast.logratio 
#>                              1.0000000                              1.0000000 
#>             pvalue.slowinskiguyer.test               pvalue.mcconwaysims.test 
#>                              0.3333333                              0.2357378 
#>              pvalue.richness.yule.test          pvalue.kafermousset.1.derived 
#>                              1.0000000                              1.0000000 
#>          pvalue.kafermousset.0.derived 
#>                              0.6224000